Add cache to subscribers count
## Problem
When the user enters www.trampardecasa.com.br the page makes a request
to `/api/subscribers` endpoint and this request response is a result of
a database query. This is not necessary and very expensive for the
database

## Solution
With [Server
component](https://nextjs.org/docs/getting-started/react-essentials#server-components)
I can control the "frequency" of rendering with caching. To make this I
export a variable called
[`revalidate`](https://nextjs.org/docs/app/api-reference/file-conventions/route-segment-config#revalidate).

The revalidate variable is part of a [Route Segment
Config](https://nextjs.org/docs/app/api-reference/file-conventions/route-segment-config)
feature.

By default, this variable is false but the time is measured in seconds