Fix: dialog close button (#127)
This PR resolves an issue related to the functionality of the dialog
close button. Previously, clicking the close button did not trigger the
desired behavior. This problem stemmed from how we managed the dialog's
open status using a state.

To fix this issue, I restructured the code by separating the dialog's
close component from the main DialogComponent. By doing so, we now have
the flexibility to pass necessary props, including the onClose function,
to the Dialogclose component.

This change allows us to ensure that the onClose function is correctly
propagated, enabling the close button to work as intended, just like in
the Link Component.

In summary, this PR enhances the dialog close button functionality by
reorganizing the code to properly incorporate the onClose function,
ensuring a seamless user experience.


![image](https://github.com/ocodista/trampar-de-casa/assets/61201131/f2687b74-28e8-47ae-9f73-ee3fa7fbdfbb)

---------
Closes #127 

Co-authored-by: Caio Borghi <caiohoborghi@gmail.com>