fix profile form (#134)
### Fix `updateSubscriber` function
This function was responsible for updating the subscribers, however when
creating the query through supabase, we were passing the "skills" field
and not "skillsId", this made it impossible to actually update the
users.

Furthermore, we were unable to visualize this error. These issues have
been resolved

### Skills field
also changes the skills component, changes the UI, and adds new options.