weekly release 2023/10/18 | fix sender-email service (#141)
In this PR, I added the vacancies sent on October 18, 2023. In addition,
we also fixed an implementation problem in the email-sending service

The way we consume does not allow us to consume small groups of
messages, send emails, and then receive more.

We realized that this strategy would no longer make sense, so we chose
to do "pooling". We do a `channel.get` and we can consume one message at
a time, this gives us more control. Additionally, we've slightly
refactored the code, removing a lot of complexity from the email sender

---------

Co-authored-by: Caio Borghi <caiohoborghi@gmail.com>