weekly release 2023/10/25 | verify if subscriber is optOut | initiate skillMatcher algorithm (#148)
This week, we used with success the `auto-email-sender` services to send
openings, initiate a `skillMatcher` algorithm, and other stuff...


### OptOut
After the `getConfirmedSubscribersRowsBlock` does not verify if
subscribers cancel the subscription. I updated this function to verify.

```ts
  const { data, error } = await supabase
    .from(Entities.Subcribers)
    .select('*')
    .eq('isConfirmed', true)
    .eq('optOut', false) // <---- if optOut column for false, return row.
```

### SkillMatcher
SkillMatcher is a robust utility module that offers skill normalization
and parsing functionalities. Its normalizer function transforms input
strings into uppercase, removes spaces, standardizes variations of 'JS',
and eliminates '.js' extensions, ensuring uniform skill representation.
The parser function utilizes a precomputed dictionary to map normalized
skills to unique IDs, enabling efficient skill categorization.

The main idea is to be able to automate the insertion of new openings as
much as possible. Imagine if it is possible to pass a URL and pass the
skills automatically, Our idea is this.

We initiate this challenge on this PR.

---------

Co-authored-by: Caio Borghi <caiohoborghi@gmail.com>