fix profile email template preview | fix english level session on profile form (#152)
In this pr I fixed the text preview of the profile email template and
found a bug on the profile form.

### Bug
When we save a profile with "Não informar" checkbox selected, the
frontend sends a "None" value for the backend. The "None" makes sense on
the frontend but this value does not exist on the database. When the
user sets "None" to the English level in the database this should be
equal to a `NULL`

---------

Co-authored-by: Caio Borghi <caiohoborghi@gmail.com>