fix roles-assigner | exit email-sender service on finish | remove unnecessary queue assert | wait drain event on `subs-to-queue` (#196)
In this PR I solved some problems, with the `roles-assigner`,
`sendToQueue` function and bug on `subs-to-queue`

### `roles-assigner` error
When run `roles-assigner`, I receive this error:

![image](https://github.com/ocodista/trampar-de-casa/assets/68869379/e5405e94-1171-491e-8dda-d3ef650655f2)

The roles-assigner not can `ack` the queue message and create an
infinite loop. To solve this, I added an error handler on
`saveSubscriberRoles` function.

### `sendToQueue`
The `await queueChannel.assertQueue(queue)` line is unnecessary and
creates a bug when sending a message to the queue.

### 'subs-to-queue`
This service on the remote queue does not wait for the drain of
messages, this causes the messages lost. we use a promise strategy for
waiting for the drain event