exit roles-renderer process | add script to create emails | add script to send emails (#201)
In this PR I fixed a bug in the roles-renderer service and added a
script for creating emails.

## roles-renderer
The roles-renderer service does not quit on finish rendering, this is a
problem for the script. To fix this I execute a `process.exit(0)` to
make the process finish.

## script
The script is necessary to make the email process more simple. and with
a script, I can register this script on crontab :)